% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDT.R
\docType{data}
\name{DDT}
\alias{DDT}
\title{DDT Data information}
\format{
A data frame with x rows and y column(s)
\describe{
  \item{RIVER}{Which river sampled from}
  \item{MILE}{Mile of river}
  \item{SPECIES}{Species of fish}
  \item{LENGTH}{Lenght of each fish}
  \item{WEIGHT}{Weight of each fish}
  \item{DDT}{Amount of DDT in each fish}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
DDT
}
\description{
Data set of river fish data used for Piecewise Regression.
   Data frame of 144 rows and 6 columns
}
\keyword{datasets}
