% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findKnot.R
\name{findKnot}
\alias{findKnot}
\title{Knot Position for Piecewise Model}
\usage{
findKnot(xlist, ylist, data, doPlot = TRUE)
}
\arguments{
\item{xlist}{String (preferred) or column number from data set referring to
the column of x values.}

\item{ylist}{String (preferred) or column number from data set referring to
the column of y values.}

\item{data}{Data set to find knot position.}

\item{doPlot}{A Boolean to have function create plots or just return list
Default is to create plots.}
}
\value{
Two plots of the AIC and \eqn{R^2} vs \eqn{x_k} are made. Also a
   named list of the following is returned (invisible):
\describe{
  \item{x_krd}{The best knot position from the discrete \eqn{R^2} calculation}
  \item{x_krc}{The best knot position from the continuous \eqn{R^2} calculation}
  \item{x_kad}{The best knot position from the discrete AIC calculation}
  \item{x_kac}{The best knot position from the continuous AIC calculation}
  \item{Rsqs}{The list of all \eqn{R^2} from the discrete calculations}
  \item{AICs}{The list of all AIC from the discrete calculations}
}
}
\description{
The function \code{findMyKnot()} will calculate the "ideal" knot
   placement for a 2-piece (one knot) piecewise regression model. It does this
   by finding the maximum \eqn{R^2} value across many different possible \eqn{x_k},
   and by finding the minimum AIC value across these possible knot positions.
   This is done in both the discrete case (where we use only the \eqn{x} values
   of the data to be used as \eqn{x_k}) and the continuous case where we let
   \eqn{x_k} potential be outside the data set.
}
\details{
This function has two "sub-functions" used to create a vector of outputs
   (either \eqn{R^2} values or AIC values for \code{findRsq} or \code{findAIC}
   respectively). The vector will be the same length as the vector of possible \eqn{x_k}
   that are passed into the function. We then use \code{map_dbl} from \code{purrr}
   to easily apply this function to our vector of \eqn{x} values. Graphs of the
   \eqn{R^2} and AIC vs \eqn{x_k} are made with a continuous line plotted. Vertical
   lines are drawn to represent the discrete "best" knot (in red) and the continuous
   version (in blue).
}
\examples{
## Example with DDT data set
print(findKnot("LENGTH", "WEIGHT", DDT))

## Example with Spruce data set
print(findKnot("BHDiameter", "Height", SPRUCE))
}
