% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRUCE.R
\docType{data}
\name{SPRUCE}
\alias{SPRUCE}
\title{Spruce Data information}
\format{
A data frame with x rows and y column(s)
\describe{
  \item{BHDiameter}{Diameter of Spruce Tree}
  \item{Height}{Height of Spruce Tree}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
SPRUCE
}
\description{
Data set of spruce tree information to be used for Piecewise Regression.
   Data frame of 36 rows and 2 columns
}
\keyword{datasets}
